mcp/requirements.txt
mlflow[databricks]>=3.1
mcp>=1.9
databricks-sdk[openai]
databricks-agents>=1.0.0
databricks-mcp
python-dotenv
asyncio
uv
